DVWA integration
If you have a working version of DVWA on your PC, the fuzzer will now be
able to login and explore the application. Only problem is that it
doesn't seem to grab any Cookies. Not sure why, but it just doesn't find
any. Not the highest priority right now.

The Commit DOES NOT contail the XAMPP server stuff